+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ New features
 
 * Add :code:`modeling.convert_unit`
 * Add :code:`modeling.set_unit`
+* Add :code:`modeling.set_property`
 
 Changes
 =======